% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brucutu.R
\name{brucutu_run}
\alias{brucutu_run}
\title{Roda script no Brucutu VI}
\usage{
brucutu_run(script_path, data_path = NULL)
}
\arguments{
\item{script_path}{caminho do script}

\item{data_path}{caminho da base de dados, se existir}
}
\description{
Roda script no Brucutu VI
}
