% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brucutu.R
\name{brucutu_unzip}
\alias{brucutu_unzip}
\title{Dezipar arquivo que está remoto}
\usage{
brucutu_unzip(zip_file)
}
\arguments{
\item{zip_file}{nome do arquivo. Precisa ter um arquivo
com esse nome em var/tmp}
}
\description{
Dezipar arquivo que está remoto
}
