% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brucutu.R
\name{brucutu_upload}
\alias{brucutu_upload}
\title{Transfere uma pasta para /var/tmp}
\usage{
brucutu_upload(path, to = "/var/tmp", home = FALSE)
}
\arguments{
\item{path}{caminho da pasta/arquivo que deseja subir}

\item{to}{pasta de destino. Recomenda-se criar uma pasta propria
em var/temp}

\item{home}{arquivo na home? Para uso interno.}
}
\description{
Transfere uma pasta para /var/tmp
}
