% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celsius_to_kelvin_doc.R
\name{celsius_to_kelvin_doc}
\alias{celsius_to_kelvin_doc}
\title{Turn Celsius into Kelvin}
\usage{
celsius_to_kelvin_doc(temp_C)
}
\arguments{
\item{temp_C}{a numeric value that indicates the temperature in Celsius degrees}
}
\value{
a numeric that indicates the equivalent temperature in kelvin of a given temperature in Celsius
}
\description{
turn the units of a given temperature in Celsius degrees to Kelvin
}
\examples{
celsius_to_kelvin(0)
}
