% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melhor_modelo.R
\name{encontrar_melhor_modelo}
\alias{encontrar_melhor_modelo}
\title{Encontra melhor modelo da simulação}
\usage{
encontrar_melhor_modelo(path)
}
\arguments{
\item{path}{pasta que contém arquivos de log dos resultados dos modelos}
}
\description{
Encontra melhor modelo partir de uma pasta com arquivos de logs. Função a ser
utilizada internamente no meu doutorado.
}
